PMO-1031: unit tests for SegmentFilter added

+++ src/groovy/org/pillarone/riskanalytics/domain/pc/constraints/SegmentPortion.groovy
@@ -24,6 +24,6 @@ class SegmentPortion implements IMultiDimensionalConstraints {
     }
 
     Class getColumnType(int column) {
-        return column == 0 ? LobMarker : BigDecimal
+        return column == 0 ? LobMarker : Double
     }
 }
